modify: renamed _json_flattener and removed FlatJSONDict
